feat(onboarding): update laptop script for 2024 (#8)
Addresses
https://app.shortcut.com/coprocure/story/21058/update-laptop-script-to-remove-installs-of-outdated-or-unnecessary-deps

A variety of changes to bring the script up to date with the newest
version of the upstream thoughtbot/laptop script. Notable changes:

- Rename default branch from master to main
- Fix handling of HOMEBREW_PREFIX for Apple Silicon & Intel Macs
- Add Rosetta 2
- Remove Heroku CLI (no longer used at Pavilion)
- Upgrade nvm to 0.39.7
- Upgrade node to 20
- Fix Homebrew install of Postgres
- Remove installation of applications using Homebrew

This last one is probably the most controversial. Basically I just think
that installing zoom via Homebrew seems kinda cursed. But I'm willing to
be outvoted (or maybe add these back in a comment).

Testing: ran this on my personal laptop and it worked a-ok.